Split KbmProfileModel types into separate files (SA1402)
